refactor(dashboard): redesign traces page to design-system v2
Replace traces-page.tsx's inline duplicate PageHeader with the shared
components/dashboard/page-header. Swap the hand-rolled <table> waterfall
list for the canonical Table/TableHeader/TableRow/TableHead/TableCell/
TableSkeleton primitives wrapped in Card, matching the already-redesigned
conversations table. Split the 443-line file into focused modules
(_traces-table, _trace-detail, _observation-row, _use-traces-data,
_trace-format) mirroring the conversations/ pattern, and reuse the shared
formatCostMicrodollars/formatDate utilities instead of duplicating local
formatters. Conversations pages required no changes — already compliant
with the v2 token/primitive conventions.
